Fix variables' type issue

+++ ogbn-mag/train_ogbn_mag.py
@@ -39,7 +39,7 @@ parser.add_argument('--n_heads', type=int, default=8,
                     help='Number of attention head')
 parser.add_argument('--n_layers', type=int, default=4,
                     help='Number of GNN layers')
-parser.add_argument('--dropout', type=int, default=0.2,
+parser.add_argument('--dropout', type=float, default=0.2,
                     help='Dropout ratio')
 parser.add_argument('--sample_depth', type=int, default=6,
                     help='How many numbers to sample the graph')
@@ -54,7 +54,7 @@ parser.add_argument('--n_batch', type=int, default=32,
                     help='Number of batch (sampled graphs) for each epoch') 
 parser.add_argument('--batch_size', type=int, default=128,
                     help='Number of output nodes for training')  
-parser.add_argument('--clip', type=int, default=1.0,
+parser.add_argument('--clip', type=float, default=1.0,
                     help='Gradient Norm Clipping') 
 
 parser.add_argument('--prev_norm', help='Whether to add layer-norm on the previous layers', action='store_true')
